extracted segment manager to its own module
