+++ src/storage/file/mod.rs
@@ -2,10 +2,12 @@ pub mod common;
 pub mod consumer_group;
 pub mod index;
 pub mod segment;
+pub mod segment_manager;
 pub mod topic_log;
 
 pub use common::SyncMode;
 pub use consumer_group::FileConsumerGroup;
 pub use index::{IndexEntry, SparseIndex};
-pub use segment::{IndexingConfig, LogSegment, SegmentManager};
+pub use segment::{IndexingConfig, LogSegment};
+pub use segment_manager::SegmentManager;
 pub use topic_log::FileTopicLog;
